requirements.txt
absl-py==1.4.0
altair==5.0.1
attrs==23.1.0
blinker==1.6.2
boto3==1.28.28
botocore==1.31.28
cachetools==5.3.1
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
contourpy==1.1.0
cycler==0.10.0
Cython==3.0.0
filelock==3.12.2
fire==0.5.0
fonttools==4.42.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
huggingface-hub==0.16.4
idna==2.10
importlib-metadata==6.8.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
lapx==0.5.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
oauthlib==3.2.2
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
packaging==23.1
pafy==0.5.5
pandas==2.0.3
Pillow==9.5.0
pip==23.2.1
protobuf==4.24.0
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybboxes==0.1.6
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytube==15.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.5.2
roboflow==1.1.3
rpds-py==0.9.2
rsa==4.9
s3transfer==0.6.2
sahi==0.11.14
scipy==1.11.1
seaborn==0.12.2
setuptools==68.1.0
shapely==2.0.1
six==1.16.0
smmap==5.0.0
streamlit==1.25.0
supervision==0.13.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
ultralytics==8.0.156
urllib3==1.26.16
validators==0.21.2
watchdog==3.0.0
Werkzeug==2.3.7
wget==3.2
wheel==0.41.1
yolov5==7.0.12
yolov8==0.0.2
zipp==3.16.2